+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v7.0.0 - 
 - **Enhancement**: Removed the FeedbackBird button from the header along with its related functionality.
 - **Fix**: Fixed issues with disappearing billing fields on woocommerce legacy and HPOS mode.
+- **New**: Added lock sections in settings/integrations pages (main plugin & addons).
 
 v6.9.12 - 2025-03-31
 - **New**: Added WhatsApp support for the Tubelight gateway.